fix datatype issues

+++ src/timeio/qc/datastream.py
@@ -26,8 +26,8 @@ __all__ = [
     "LocalStream",
 ]
 
-""" 
-This module provides a convenient abstraction for retrieving and 
+"""
+This module provides a convenient abstraction for retrieving and
 storing datastreams from/to the users observation database.
 """
 
@@ -135,14 +135,14 @@ class DatastreamSTA:
         query = sql.SQL(
             """
             select "RESULT_TIME", "RESULT_TYPE", "RESULT_NUMBER", "RESULT_STRING",
-                "RESULT_JSON", "RESULT_BOOLEAN", "RESULT_QUALITY",  l.datastream_id 
-            from "OBSERVATIONS" o 
-            join public.sms_datastream_link l on o."DATASTREAM_ID" = l.device_property_id 
-            where o."DATASTREAM_ID" = %s 
-              and o."RESULT_TIME" >= %s 
-              and o."RESULT_TIME" <= %s 
-            order by o."RESULT_TIME" desc 
-            limit %s 
+                "RESULT_JSON", "RESULT_BOOLEAN", "RESULT_QUALITY",  l.datastream_id
+            from "OBSERVATIONS" o
+            join public.sms_datastream_link l on o."DATASTREAM_ID" = l.device_property_id
+            where o."DATASTREAM_ID" = %s
+              and o."RESULT_TIME" >= %s
+              and o."RESULT_TIME" <= %s
+            order by o."RESULT_TIME" desc
+            limit %s
             """
         )
 
@@ -224,11 +224,11 @@ class DatastreamSTA:
         self,
     ) -> tuple[TimestampT, TimestampT] | tuple[None, None]:
         """Returns (earliest, latest) timestamp of data that was never seen by QC."""
-        query = """ 
+        query = """
            select o."RESULT_TIME" from "OBSERVATIONS" o
-           where o."DATASTREAM_ID" = %s 
+           where o."DATASTREAM_ID" = %s
              and (o."RESULT_QUALITY" is null or o."RESULT_QUALITY" = 'null'::jsonb)
-           order by "RESULT_TIME" {order} 
+           order by "RESULT_TIME" {order}
            limit 1
         """
 
@@ -292,7 +292,7 @@ class DatastreamSTA:
             return chunk
 
         context_start = self._fetch_context(date_start, context_window)
-        return self._data.loc[context_start:date_end, self._columns]
+        return self._data.loc[context_start:date_end, self._columns].astype(float)
 
     def __repr__(self):
         klass = self.__class__.__name__
@@ -424,7 +424,7 @@ class ProductStream(Datastream):
         # the same query on different tables.
         query = sql.SQL(
             """
-            select o.result_time, o.result_type, o.result_number, o.result_string, 
+            select o.result_time, o.result_type, o.result_number, o.result_string,
                    o.result_json, o.result_boolean, o.result_quality, o.datastream_id
             from observation o
             join datastream d on o.datastream_id = d.id
